+++ examples/custom_shape.rb
@@ -1,11 +1,11 @@
 require 'glimmer-dsl-libui'
 
-class BasicShape
+class CustomShape
   include Glimmer::LibUI::Application
   
   body {
     window {
-      title 'Hello, Shape!'
+      title 'Custom Shape'
       content_size 200, 225
     
       @area = area {
@@ -13,19 +13,22 @@ class BasicShape
           fill :white
         }
         
-        15.times do |n|
+        7.times do |n|
           x_location = (rand*125).to_i%200 + (rand*15).to_i
           y_location = (rand*125).to_i%200 + (rand*15).to_i
-          shape_color = [rand*255, rand*255, rand*255]
-          shape_size = 10+n
+          shape_color = [rand*125 + 130, rand*125 + 130, rand*125 + 130]
+          shape_size = 20+n
 
           cube(location_x: x_location,
                location_y: y_location,
                rectangle_width: shape_size*2,
                rectangle_height: shape_size,
-               cube_height: shape_size*3,
+               cube_height: shape_size*2,
                background_color: shape_color,
-               line_thickness: 2)
+               line_thickness: 2) {
+            # TODO support listeners
+            # TODO support extra properties
+          }
         end
       }
     }
@@ -48,28 +51,43 @@ class BasicShape
     foreground_color = [0, 0, 0, thickness: line_thickness]
     
     shape(location_x, location_y) { |the_shape|
-      polygon(0, cube_height + rectangle_height / 2.0, rectangle_width / 2.0, cube_height, rectangle_width, cube_height + rectangle_height / 2.0, rectangle_width / 2.0, cube_height + rectangle_height) {
+      polygon(0, cube_height + rectangle_height / 2.0,
+              rectangle_width / 2.0, cube_height,
+              rectangle_width, cube_height + rectangle_height / 2.0,
+              rectangle_width / 2.0, cube_height + rectangle_height) {
         fill background_color
         stroke foreground_color
       }
       rectangle(0, rectangle_height / 2.0, rectangle_width, cube_height) {
         fill background_color
       }
-      polyline(0, rectangle_height / 2.0 + cube_height, 0, rectangle_height / 2.0, rectangle_width, rectangle_height / 2.0, rectangle_width, rectangle_height / 2.0 + cube_height) {
+      polyline(0, rectangle_height / 2.0 + cube_height,
+               0, rectangle_height / 2.0,
+               rectangle_width, rectangle_height / 2.0,
+               rectangle_width, rectangle_height / 2.0 + cube_height) {
         stroke foreground_color
       }
-      polygon(0, rectangle_height / 2.0, rectangle_width / 2.0, 0, rectangle_width, rectangle_height / 2.0, rectangle_width / 2.0, rectangle_height) {
+      polygon(0, rectangle_height / 2.0,
+              rectangle_width / 2.0, 0,
+              rectangle_width, rectangle_height / 2.0,
+              rectangle_width / 2.0, rectangle_height) {
         fill background_color
         stroke foreground_color
       }
-      line(rectangle_width / 2.0, cube_height + rectangle_height, rectangle_width / 2.0, rectangle_height) {
+      line(rectangle_width / 2.0, cube_height + rectangle_height,
+           rectangle_width / 2.0, rectangle_height) {
         stroke foreground_color
       }
       
       content_block&.call(the_shape)
+      
+      # TODO do more with this
+      on_mouse_up do |area_mouse_event|
+        puts 'Mouse Up'
+      end
     }
   end
 end
 
-BasicShape.launch
+CustomShape.launch