+++ lib/glimmer/libui/shape/arc.rb
@@ -51,15 +51,15 @@ module Glimmer
         end
         
         def perfect_shape
-          perfect_shape_dependencies = [x_center, y_center, radius, start_angle, sweep, is_negative]
+          perfect_shape_dependencies = [absolute_x_center, absolute_y_center, radius, start_angle, sweep, is_negative]
           if perfect_shape_dependencies != @perfect_shape_dependencies
-            x_center, y_center, radius, start_angle, sweep, is_negative = @perfect_shape_dependencies = perfect_shape_dependencies
+            absolute_x_center, absolute_y_center, radius, start_angle, sweep, is_negative = @perfect_shape_dependencies = perfect_shape_dependencies
             sign = is_negative ? 1 : -1
             start = is_negative ? (360 - start_angle) : -1*start_angle
             extent = is_negative ? (360 - sweep) : -1*sweep
             @perfect_shape = PerfectShape::Arc.new(
               type: :open,
-              center_x: x_center, center_y: y_center,
+              center_x: absolute_x_center, center_y: absolute_y_center,
               radius_x: radius, radius_y: radius,
               start: start, extent: extent
             )