+++ lib/glimmer/libui/shape/circle.rb
@@ -51,10 +51,10 @@ module Glimmer
         end
         
         def perfect_shape
-          perfect_shape_dependencies = [x_center, y_center, radius]
+          perfect_shape_dependencies = [absolute_x_center, absolute_y_center, radius]
           if perfect_shape_dependencies != @perfect_shape_dependencies
-            x_center, y_center, radius = @perfect_shape_dependencies = perfect_shape_dependencies
-            @perfect_shape = PerfectShape::Circle.new(center_x: x_center, center_y: y_center, radius: radius)
+            absolute_x_center, absolute_y_center, radius = @perfect_shape_dependencies = perfect_shape_dependencies
+            @perfect_shape = PerfectShape::Circle.new(center_x: absolute_x_center, center_y: absolute_y_center, radius: radius)
           end
           @perfect_shape
         end