+++ lib/glimmer/libui/shape/polyline.rb
@@ -45,11 +45,11 @@ module Glimmer
         end
         
         def perfect_shape
-          perfect_shape_dependencies = PerfectShape::MultiPoint.normalize_point_array(point_array)
+          perfect_shape_dependencies = absolute_point_array
           if perfect_shape_dependencies != @perfect_shape_dependencies
-            point_array = @perfect_shape_dependencies = perfect_shape_dependencies
-            path_shapes = [point_array.first]
-            path_shapes += point_array.drop(1).map { |point| PerfectShape::Line.new(points: [point]) }
+            absolute_point_array = @perfect_shape_dependencies = perfect_shape_dependencies
+            path_shapes = [absolute_point_array.first]
+            path_shapes += absolute_point_array.drop(1).map { |point| PerfectShape::Line.new(points: [point]) }
             @perfect_shape = PerfectShape::Path.new(closed: false, shapes: path_shapes)
           end
           @perfect_shape